# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Interface API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Workflow {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  icon: string;
  threadCount?: number;
  createdAt: string;
}

export interface WorkflowInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  color: string;
  icon: string;
}

export interface WorkflowUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  color?: string;
  icon?: string;
}

export type ThreadStatus = typeof ThreadStatus[keyof typeof ThreadStatus];


export const ThreadStatus = {
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Thread {
  id: number;
  title: string;
  /** @nullable */
  workflowId?: number | null;
  /** @nullable */
  workflowName?: string | null;
  /** @nullable */
  workflowColor?: string | null;
  status: ThreadStatus;
  pinned: boolean;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ThreadInput {
  /** @minLength 1 */
  title: string;
  workflowId?: number;
}

export type ThreadUpdateStatus = typeof ThreadUpdateStatus[keyof typeof ThreadUpdateStatus];


export const ThreadUpdateStatus = {
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface ThreadUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  workflowId?: number | null;
  status?: ThreadUpdateStatus;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  agent: 'agent',
} as const;

export interface Message {
  id: number;
  threadId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  telegramMessageId?: number | null;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export type DashboardSummaryWorkflowBreakdownItem = {
  workflowId: number;
  workflowName: string;
  color: string;
  threadCount: number;
};

export interface DashboardSummary {
  totalThreads: number;
  activeThreads: number;
  totalMessages: number;
  workflowBreakdown: DashboardSummaryWorkflowBreakdownItem[];
}

export interface Settings {
  /** @nullable */
  telegramBotToken: string | null;
  /** @nullable */
  telegramChatId: string | null;
  configured: boolean;
}

export interface SettingsUpdate {
  telegramBotToken?: string;
  telegramChatId?: string;
}

export type ListThreadsParams = {
workflowId?: number;
};