# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hermes Agent Interface API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: threads
    description: Conversation thread operations
  - name: messages
    description: Message operations
  - name: workflows
    description: Workflow category operations
  - name: settings
    description: Settings operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List all workflow categories
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"

  /workflows/{id}:
    patch:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          description: Not found
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /threads:
    get:
      operationId: listThreads
      tags: [threads]
      summary: List all threads, optionally filtered by workflow
      parameters:
        - name: workflowId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Thread"
    post:
      operationId: createThread
      tags: [threads]
      summary: Create a new conversation thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"

  /threads/{id}:
    get:
      operationId: getThread
      tags: [threads]
      summary: Get a thread by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
        "404":
          description: Not found
    patch:
      operationId: updateThread
      tags: [threads]
      summary: Update a thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
        "404":
          description: Not found
    delete:
      operationId: deleteThread
      tags: [threads]
      summary: Delete a thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /threads/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "404":
          description: Thread not found
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a thread (relays to Telegram and stores response)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Thread not found

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [threads]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /threads/{id}/pin:
    patch:
      operationId: pinThread
      tags: [threads]
      summary: Toggle pin state of a thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
        "404":
          description: Not found

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get app settings (Telegram config)
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Workflow:
      type: object
      required: [id, name, color, icon, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
        icon:
          type: string
        threadCount:
          type: integer
        createdAt:
          type: string

    WorkflowInput:
      type: object
      required: [name, color, icon]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
        icon:
          type: string

    WorkflowUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
        icon:
          type: string

    Thread:
      type: object
      required: [id, title, status, pinned, messageCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        workflowId:
          type: ["integer", "null"]
        workflowName:
          type: ["string", "null"]
        workflowColor:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, archived]
        pinned:
          type: boolean
        messageCount:
          type: integer
        lastMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ThreadInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        workflowId:
          type: integer

    ThreadUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        workflowId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, completed, archived]

    Message:
      type: object
      required: [id, threadId, role, content, createdAt]
      properties:
        id:
          type: integer
        threadId:
          type: integer
        role:
          type: string
          enum: [user, agent]
        content:
          type: string
        telegramMessageId:
          type: ["integer", "null"]
        createdAt:
          type: string

    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    DashboardSummary:
      type: object
      required: [totalThreads, activeThreads, totalMessages, workflowBreakdown]
      properties:
        totalThreads:
          type: integer
        activeThreads:
          type: integer
        totalMessages:
          type: integer
        workflowBreakdown:
          type: array
          items:
            type: object
            required: [workflowId, workflowName, color, threadCount]
            properties:
              workflowId:
                type: integer
              workflowName:
                type: string
              color:
                type: string
              threadCount:
                type: integer

    Settings:
      type: object
      required: [telegramBotToken, telegramChatId, configured]
      properties:
        telegramBotToken:
          type: ["string", "null"]
        telegramChatId:
          type: ["string", "null"]
        configured:
          type: boolean

    SettingsUpdate:
      type: object
      properties:
        telegramBotToken:
          type: string
        telegramChatId:
          type: string