# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Interface API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryWorkflowBreakdownItem } from './dashboardSummaryWorkflowBreakdownItem';

export interface DashboardSummary {
  totalThreads: number;
  activeThreads: number;
  totalMessages: number;
  workflowBreakdown: DashboardSummaryWorkflowBreakdownItem[];
}