# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Interface API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  /** @nullable */
  telegramBotToken: string | null;
  /** @nullable */
  telegramChatId: string | null;
  configured: boolean;
}