# lib/api-zod/src/generated/types/thread.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Interface API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreadStatus } from './threadStatus';

export interface Thread {
  id: number;
  title: string;
  /** @nullable */
  workflowId?: number | null;
  /** @nullable */
  workflowName?: string | null;
  /** @nullable */
  workflowColor?: string | null;
  status: ThreadStatus;
  pinned: boolean;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}