# lib/api-zod/src/generated/types/threadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Interface API
 * OpenAPI spec version: 0.1.0
 */

export interface ThreadInput {
  /** @minLength 1 */
  title: string;
  workflowId?: number;
}