# lib/api-zod/src/generated/types/threadUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Interface API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreadUpdateStatus } from './threadUpdateStatus';

export interface ThreadUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  workflowId?: number | null;
  status?: ThreadUpdateStatus;
}