# lib/api-zod/src/generated/types/workflow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Interface API
 * OpenAPI spec version: 0.1.0
 */

export interface Workflow {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  icon: string;
  threadCount?: number;
  createdAt: string;
}