# lib/api-zod/src/generated/types/workflowUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hermes Agent Interface API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkflowUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  color?: string;
  icon?: string;
}